Extract connect and transcoding routes into separate classes, split connect and transcoding into packages
